% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_count_linear_weight.R
\name{compute_count_linear_weight}
\alias{compute_count_linear_weight}
\title{Compute count linear weight}
\usage{
compute_count_linear_weight(pitch, event, base_out_run_exp)
}
\arguments{
\item{pitch}{dataframe of pitch data from \code{\link{extract_season}}}

\item{event}{dataframe of event data from \code{\link{extract_season}}}

\item{base_out_run_exp}{dataframe of base-out run expectancy from
\code{\link{compute_base_out_run_exp}}}
}
\value{
a dataframe of \code{linear_weight} indexed by \code{balls} and \code{strikes}
}
\description{
Compute the average change in base-out run expectancy for all plate appearances that run
through each count.
}
