% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pitch_value.R
\name{compute_pitch_value}
\alias{compute_pitch_value}
\title{Compute pitch value}
\usage{
compute_pitch_value(pitch_pred, count_value)
}
\arguments{
\item{pitch_pred}{dataframe of predictions from a pitch outcome model}

\item{count_value}{dataframe of count_value from \code{\link{compute_count_value}}}
}
\description{
Using pitch outcome model predictions, compute run value (expected linear weight).
}
