% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_hit_outcome_model.R
\docType{data}
\name{config_hit_outcome_xgb}
\alias{config_hit_outcome_xgb}
\title{Configuration for hit outcome XGBoost model}
\format{
An object of class \code{list} of length 4.
}
\usage{
config_hit_outcome_xgb
}
\description{
Use this object to specify which values to use for each tuning parameter.
We have a separate script for determining optimal parameters.
}
\keyword{datasets}
