% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pitch_outcome_model.R
\docType{data}
\name{config_pitch_outcome_xgb}
\alias{config_pitch_outcome_xgb}
\title{Configuration for pitch outcome XGBoost models}
\format{
An object of class \code{list} of length 13.
}
\usage{
config_pitch_outcome_xgb
}
\description{
Configuration for pitch outcome XGBoost models
}
\keyword{datasets}
