% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_savant_data.R
\name{download_savant_data}
\alias{download_savant_data}
\title{Download Statcast data from Baseball Savant}
\usage{
download_savant_data(start_date, end_date, verbose = FALSE)
}
\arguments{
\item{start_date}{first date included in the download}

\item{end_date}{last date included in the download}

\item{verbose}{logical, should progress be printed?, defaults to FALSE}
}
\value{
a dataframe with 94 columns and one row per pitch, with all available data
}
\description{
Loop over five days at a time to download data from the Statcast search API at Baseball Savant.
This includes swing tracking data not available through the MLB statsapi.
}
