% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_population_variance.R
\name{estimate_population_variance}
\alias{estimate_population_variance}
\title{Estimate population variance}
\usage{
estimate_population_variance(
  observed_value,
  noise_variance,
  population_mean = weighted.mean(observed_value, w = 1/noise_variance, na.rm = TRUE),
  max_iterations = 10000,
  tolerance = 1e-07 * var(observed_value)
)
}
\arguments{
\item{observed_value}{vector of observed performance}

\item{noise_variance}{vector of estimated noise variance}

\item{population_mean}{scalar, population mean for \code{observed_value}}

\item{max_iterations}{integer, optimization stops after \code{max_iterations} is reached}

\item{tolerance}{scalar, optimization stops after movement in estimated population variance is
less than tolerance}
}
\value{
a scalar estimate of population variance
}
\description{
Estimate the population variance in true talent for a performance metric given a vector of
observed values and associated noise variances.
}
