% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_season.R
\name{extract_season}
\alias{extract_season}
\title{Extract season}
\usage{
extract_season(year, level = c("mlb", "aaa"), cl = NULL)
}
\arguments{
\item{year}{an integer specifying the year to extract}

\item{level}{character string, "mlb" (default) or "aaa"}

\item{cl}{optional cluster object for parallel computation, default is NULL (not parallel)}
}
\value{
a list of two dataframes: \code{event} and \code{pitch}
}
\description{
Extract all regular season MLB games in a specified year.
}
