% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_model_features.R
\name{get_outcome_model_features}
\alias{get_outcome_model_features}
\title{Get features for pitch outcome model}
\usage{
get_outcome_model_features(data)
}
\arguments{
\item{data}{dataframe with the following columns: ax, ay, az, bx, by, bz, cx, cy, cz}
}
\value{
a dataframe with the following columns:
plate_x, plate_z, plate_vx, plate_vy, plate_vz, ax, az, az, extension
}
\description{
We elect to parameterize each pitch based on its characteristics at the plate because these are
most directly relevant to the batter. The one exception is extension, which we include because
it has no analog at the plate (plate_y is always 17/12).
}
