% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trackman_metrics.R
\name{get_trackman_metrics}
\alias{get_trackman_metrics}
\title{Get TrackMan metrics for pitch data}
\usage{
get_trackman_metrics(data)
}
\arguments{
\item{data}{dataframe with the following columns: ax, ay, az, bx, by, bz, cx, cy, cz}
}
\value{
a dataframe with the following columns:
release_x, release_y, release_z, plate_x, plate_z, horz_break, vert_break, induced_vert_break
}
\description{
Use the quadratic coefficients describing the path of the pitch to calculate TrackMan metrics
}
