% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2d_function.R
\name{plot_2d_function}
\alias{plot_2d_function}
\title{Visualize a function of two variables}
\usage{
plot_2d_function(x, y, z, z_min = min(z), z_max = max(z), pch = 15, ...)
}
\arguments{
\item{x}{the first input to the function (x-axis)}

\item{y}{the second input to the function (y-axis)}

\item{z}{the output of the function (color)}

\item{z_min}{the value of z corresponding to the deepest blue, must be <= min(z)}

\item{z_max}{the value of z corresponding to the deepest red, must be >= max(z)}

\item{pch}{plotting character, default is 15 (solid square)}

\item{...}{additional arguments passed to \code{plot()}}
}
\description{
This function is useful for plotting model fits for supervised regression models with two
features. It works best if x and y are laid out on a grid, rather than reflecting data samples.
}
