% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch_outcome_model.R
\name{predict.pitch_outcome_model}
\alias{predict.pitch_outcome_model}
\title{Extract predictions from a fitted pitch outcome model}
\usage{
\method{predict}{pitch_outcome_model}(object, newdata, ...)
}
\arguments{
\item{object}{a fitted "pitch_outcome_model" object}

\item{newdata}{a dataframe of pitch data for which to make predictions}
}
\value{
a dataframe of predictions with columns swing, hbp, strike, contact and fair
}
\description{
Extract predictions from a fitted pitch outcome model
}
