% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pitches.R
\name{sim_pitches}
\alias{sim_pitches}
\title{Simulate pitches according to pitch distribution model}
\usage{
sim_pitches(pitch_distribution_model, pitcher_id, n, context)
}
\arguments{
\item{pitcher_id}{integer, ID of pitcher to simulate (must be length-1)}

\item{n}{integer, number of pitches to simulate}

\item{context}{a dataframe with the following columns:
\code{bsh_num}, \code{bat_side}, \code{pre_balls}, \code{pre_strikes}, \code{strike_zone_top}, \code{strike_zone_bottom}}

\item{model}{a fitted "pitch_distrib_model" object}
}
\value{
a dataframe of simulated pitch characteristics
}
\description{
For a single pitcher, we draw pitch characteristic z-scores based on their estimated Rho matrix,
and then we re-center/re-scale those simulated z-scores based on the context (which is itself
drawn with replacement from the dataframe provided) and the pitcher's estimated parameters.
}
