% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pitch_score.R
\name{simulate_pitch_score}
\alias{simulate_pitch_score}
\title{Simulate pitch score}
\usage{
simulate_pitch_score(
  pitcher_id,
  n,
  context,
  pitch_distrib_model,
  pitch_outcome_model
)
}
\arguments{
\item{pitcher_id}{integer, ID of pitcher to simulate (must be length-1)}

\item{n}{integer, number of pitches to simulate}

\item{context}{a dataframe with the following columns:
\code{bat_side}, \code{pre_balls}, \code{pre_strikes}, \code{strike_zone_top}, \code{strike_zone_bottom}}

\item{pitch_distrib_model}{a fitted "pitch_distrib_model" object}

\item{pitch_outcome_model}{a fitted "pitch_outcome_model" object}
}
\value{
a scalar pitch score
}
\description{
Draw samples from the estimated pitch distribution for a specified pitcher, and run them through
the pitch outcome model to report the average predicted pitch value.
}
