% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pitches.R
\name{simulate_pitches}
\alias{simulate_pitches}
\title{Simulate pitches according to pitch distribution model}
\usage{
simulate_pitches(model, pitcher_id, n, context)
}
\arguments{
\item{model}{a fitted "pitch_distrib_model" object}

\item{pitcher_id}{integer, ID of pitcher to simulate (must be length-1)}

\item{n}{integer, number of pitches to simulate}

\item{context}{a dataframe with the following columns:
\code{bat_side}, \code{balls}, \code{strikes}, \code{strike_zone_top}, \code{strike_zone_bottom}}
}
\value{
a dataframe of simulated pitch characteristics
}
\description{
For a single pitcher, we draw pitch characteristic z-scores based on their estimated Rho matrix,
and then we re-center/re-scale those simulated z-scores based on the context (which is itself
drawn with replacement from the dataframe provided) and the pitcher's estimated parameters.
}
