% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_base_out_state.R
\name{track_base_out_state}
\alias{track_base_out_state}
\title{Track base-out state}
\usage{
track_base_out_state(event_data)
}
\arguments{
\item{event_data}{This is the liveData$plays$allPlays section of the JSON from a game's GUMBO feed}
}
\value{
a tibble with the following columns:
event_index, mid_event_outs, mid_event_runs,
pre_event_runner_1b_id, pre_event_runner_2b_id, pre_event_runner_3b_id, pre_event_out,
pre_play_runner_1b_id, pre_play_runner_2b_id, pre_play_runner_3b_id, pre_play_out,
post_runner_1b_id, post_runner_2b_id, post_runner_3b_id, post_out
}
\description{
Post-event base-out state is given to us for free. We get pre-event base-out state based on the
post-event base-out state of the previous event. However, we also need to get the pre-play
base-out state, which describes the base-out state prior to the final pitch of the event.
To do so, we start with the pre-event base-out state and account for mid-event runner movement.
}
