% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pitch_distrib_model.R
\name{train_pitch_distrib_model}
\alias{train_pitch_distrib_model}
\title{Estimate pitch distribution model via MAP parameters}
\usage{
train_pitch_distrib_model(
  data,
  version = c("complete", "conditional"),
  complete_model = NULL,
  pitch_char_vec = c("ax", "bx", "cx", "ay", "by", "cy", "az", "bz", "cz"),
  ...
)
}
\arguments{
\item{data}{dataframe of pitch and event data}

\item{version}{character string, "complete" or "conditional". The complete model assumes league
hyperparameters are unknown (takes longer), whereas the conditional model estimates pitcher
parameters conditional on previously estimated league hyperparameters.}

\item{complete_model}{a fitted "pitch_distrib_model" object required if version == "conditional",
otherwise ignored}

\item{pitch_char_vec}{character vector of pitch characteristics to model}

\item{...}{additional arguments for for cmdstanr optimize method}
}
\value{
a fitted "pitch_distrib_model" object
}
\description{
We use cmdstanr to estimate MAP parameters for a hierarchical Bayesian model for pitch
characteristic means, standard deviations, and covariance matrices.
}
