% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch_outcome_model.R
\name{train_pitch_outcome_model}
\alias{train_pitch_outcome_model}
\title{Train pitch outcome model}
\usage{
train_pitch_outcome_model(data)
}
\arguments{
\item{data}{a dataframe of pitch data with which to train the pitch outcome model}
}
\value{
a fitted "pitch_outcome_model" object
}
\description{
Train models to predict the outcome tree of a pitch: probability of swing; probability of hbp
given no swing; probability of strike given no swing and no hbp; probability of contact given
swing; and probability of fair given contact.
}
