% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pitch_outcome_model.R
\name{train_pitch_outcome_xgb}
\alias{train_pitch_outcome_xgb}
\title{Train a pitch outcome sub-model using xgboost}
\usage{
train_pitch_outcome_xgb(
  data_subset,
  features,
  tune,
  label = c("swing", "hbp", "strike", "contact", "fair", "hit", "value"),
  verbose = 0,
  ...
)
}
\arguments{
\item{data_subset}{a dataframe of pitch data subsetted to pitches of interest}

\item{features}{a character vector of column names from \code{data_subset} to be used as features}

\item{tune}{logical, if true, tune the hyperparameters instead of fitting the model}

\item{label}{a character string, one of "swing", "hbp", "strike", "contact", "fair", "hit",
used to determine which tuning parmeters to use from \code{\link{config_pitch_outcome_xgb}}}

\item{verbose}{argument passed directly to \code{xgboost::xgboost}, defaults to silent}

\item{...}{additional parameters to pass to \code{xgboost::xgboost}}
}
\value{
a fitted \code{xgboost::xgboost} model
}
\description{
This function is a wrapper for xgboost to be used for training the sub-components of the pitch
outcome model.
}
