% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_xgb_parallel.R
\name{tune_xgb_parallel}
\alias{tune_xgb_parallel}
\title{Tune an XGBoost model while leveraging parallel computation}
\usage{
tune_xgb_parallel(
  covariate_matrix,
  label,
  params_list,
  nrounds,
  response = c("gaussian", "binomial")
)
}
\arguments{
\item{covariate_matrix}{a matrix of predictor variables to pass to \code{xgboost::xgb.DMatrix}}

\item{label}{a vector response variable to pass to \code{xgboost::xgb.DMatrix}}

\item{params_list}{a list of lists of parameters (each list of parameters is evaluated by CV)}

\item{nrounds}{integer number of boosting trees to pass to \code{xgboost::xgb.cv}}

\item{response}{character, must be "gaussian" or "binomial"}
}
\value{
a dataframe of estimated CV errors indexed by parameters
}
\description{
This function is primarily a wrapper for \code{xgboost::xgb.cv}, which calculates CV errors for
provided parameter combinations. It leverages parallelization through
\code{future.apply::future_lapply}.
}
