% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pitch_distrib_model.R
\name{train_pitch_distrib_model}
\alias{train_pitch_distrib_model}
\title{Estimate pitch distribution model via MAP parameters}
\usage{
train_pitch_distrib_model(
  pitch,
  event,
  pt = c("FF", "SI", "FC", "SL", "CU", "KC", "CH", "FS"),
  pitch_char_vec = c("ax", "bx", "cx", "ay", "by", "cy", "az", "bz", "cz"),
  ...
)
}
\arguments{
\item{pitch}{dataframe of pitch data from \code{\link{extract_season}}}

\item{event}{dataframe of event data from \code{\link{extract_season}}}

\item{pt}{character, abbreviated pitch type for which to fit model}

\item{pitch_char_vec}{character vector of pitch characteristics to model}

\item{...}{additional arguments for for cmdstanr optimize method}
}
\value{
a fitted "pitch_distrib_model" object
}
\description{
We use cmdstanr to estimate MAP parameters for a hierarchical Bayesian model for pitch
characteristic means, standard deviations, and covariance matrices.
}
