% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pitch_outcome_model.R
\name{train_pitch_outcome_model}
\alias{train_pitch_outcome_model}
\title{Train pitch outcome model}
\usage{
train_pitch_outcome_model(pitch)
}
\arguments{
\item{pitch}{dataframe of pitch data from \code{\link{extract_season}}}
}
\value{
a fitted "pitch_outcome_model" object
}
\description{
Train models to predict the outcome tree of a pitch: probability of swing; probability of hbp
given no swing; probability of strike given no swing and no hbp; probability of contact given
swing; and probability of fair given contact.
}
