% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pitch_outcome_model.R
\name{train_pitch_outcome_model}
\alias{train_pitch_outcome_model}
\title{Train pitch outcome model}
\usage{
train_pitch_outcome_model(
  pitch,
  count_value,
  models_to_fit = c("pitch_swing", "pitch_hbp", "pitch_strike", "pitch_contact",
    "pitch_fair", "pitch_hit", "pitch_value"),
  stuff_only = FALSE,
  tune = FALSE
)
}
\arguments{
\item{pitch}{dataframe of pitch data from \code{\link{extract_season}}}

\item{count_value}{dataframe of count_value from \code{\link{compute_count_value}},
not used but stashed in the model object for pitch value prediction}

\item{models_to_fit}{character vector specifying which component models to fit}

\item{stuff_only}{logical, fit model with only "Stuff" features (no pitch location)?}

\item{tune}{logical, if true, tune the hyperparameters instead of fitting the model}
}
\value{
a fitted "pitch_outcome_model" object
}
\description{
Train models to predict the outcome tree of a pitch: probability of swing; probability of hbp
given no swing; probability of strike given no swing and no hbp; probability of contact given
swing; and probability of fair given contact.
}
