% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_pitch_distrib.R
\name{visualize_pitch_distrib}
\alias{visualize_pitch_distrib}
\title{Visualize pitch distribution}
\usage{
visualize_pitch_distrib(
  model,
  pitcher_id,
  data = NULL,
  n = 10000,
  bat_side = c("L", "R"),
  pre_balls = 0,
  pre_strikes = 0,
  strike_zone_top = 3.4,
  strike_zone_bottom = 1.6,
  ...
)
}
\arguments{
\item{model}{a fitted "pitch_distrib_model" object}

\item{pitcher_id}{integer, specifies which pitcher to visualize}

\item{data}{optional dataframe of individual pitches to plot as circles}

\item{n}{number of pitches to sample from the estimated distribution}

\item{bat_side}{character, must be "L" or "R"}

\item{pre_balls}{integer, number of balls for context}

\item{pre_strikes}{integer, number of strikes for context}

\item{strike_zone_top}{numeric, top of strike zone for context}

\item{strike_zone_bottom}{numeric, bottom of strike zone for context}
}
\value{
invisible, but a plot is produced
}
\description{
Plot a heatmap of the estimated pitch distribution for a specified pitcher and context
}
